created the fail

+++ src/component/DeleteCategory/DeleteCategory.js
@@ -12,6 +12,13 @@ export default function DeleteCategory ({ category }) {
   }
 
   function onDeleteClick() {
+    const count = Object.values(products).filter(product => product.category === category.id).length;
+
+    if (count > 0) {
+      alert("This category has existing products. Please delete them before deleting a category.");
+
+      return;
+    }
     deleteDoc(doc(db, "categories", category.id));
   }
